Refactor: remove obsolete GitBlameLines from html data
Since it's already contained in MatchingPath. Also fix tests,
fix bazel files, extract git.go from stats.go
