--- robots/cmd/test-label-analyzer/cmd/testdata/BUILD.bazel
@@ -1,7 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_test")
-
-go_test(
-    name = "go_default_test",
-    srcs = ["simple_test.go"],
-    deps = ["@com_github_onsi_ginkgo_v2//:go_default_library"],
-)
